+++ src/test/java/com/maxmind/minfraud/response/DeviceTest.java
@@ -3,6 +3,8 @@ package com.maxmind.minfraud.response;
 import com.fasterxml.jackson.jr.ob.JSON;
 import org.junit.Test;
 
+import java.util.UUID;
+
 import static org.junit.Assert.assertEquals;
 
 public class DeviceTest extends AbstractOutputTest {
@@ -19,6 +21,6 @@ public class DeviceTest extends AbstractOutputTest {
                         .finish()
         );
 
-        assertEquals("C8D3BE1A-BE26-11E5-8C50-1B575C37265F", device.getId());
+        assertEquals(UUID.fromString("C8D3BE1A-BE26-11E5-8C50-1B575C37265F"), device.getId());
     }
 }